Fix problems on Expense insertion

+++ core/models.py
@@ -25,7 +25,7 @@ class ExpenseType(AutoTimestamps):
 class Expense(AutoTimestamps):
     description = models.TextField("Description", blank=True, null=True)
     unit_price = models.DecimalField("Unit Price", max_digits=10, decimal_places=2)
-    quantity = models.FloatField("Quantity")
+    quantity = models.DecimalField("Quantity", max_digits=10, decimal_places=2)
     shipping_price = models.DecimalField("Shipping Price", max_digits=10, decimal_places=2)
     type = models.ForeignKey(ExpenseType, verbose_name="Expense Type", on_delete=models.CASCADE)
     owner = models.ForeignKey(User, verbose_name="Owner", on_delete=models.CASCADE)
